PL-8059 Unexpected process exit when using subdesigns
(cherry picked from commit 7ef045a525147109a172f1d1dc889f7e2fd2e5a1)

# Conflicts:
#	modules/core/src/com/haulmont/workflow/core/app/design/DesignerServiceBean.java
#	modules/core/src/com/haulmont/workflow/core/app/design/modules/SubDesignModule.java

+++ modules/core/src/com/haulmont/workflow/core/app/design/DesignerWorker.java
@@ -297,10 +297,12 @@ public class DesignerWorker implements DesignerWorkerAPI {
         for (DesignScript script : design.getScripts()) {
             em.persist(script);
         }
-        if (CollectionUtils.isNotEmpty(design.getDesignProcessVariables()))
-            for (DesignProcessVariable variable : design.getDesignProcessVariables()) {
-                em.persist(variable);
-            }
+        if (design.getDesignProcessVariables() != null) {
+            if (CollectionUtils.isNotEmpty(design.getDesignProcessVariables()))
+                for (DesignProcessVariable variable : design.getDesignProcessVariables()) {
+                    em.persist(variable);
+                }
+        }
         return design;
     }
 
@@ -312,8 +314,10 @@ public class DesignerWorker implements DesignerWorkerAPI {
         for (DesignScript script : existsDesign.getScripts()) {
             em.remove(script);
         }
-        for (DesignProcessVariable variable : existsDesign.getDesignProcessVariables()) {
-            em.remove(variable);
+        if (existsDesign.getDesignProcessVariables() != null) {
+            for (DesignProcessVariable variable : existsDesign.getDesignProcessVariables()) {
+                em.remove(variable);
+            }
         }
         for (DesignFile designFile : existsDesign.getDesignFiles()) {
             em.remove(designFile);
